drv: Separate the project into lis2mdl (original) and lsm6ds drivers
