+++ projects/01drv_lsm6ds/01drv_lsm6ds.c
@@ -1,10 +1,10 @@
 /**
- * @file 01drv_lis2mdl.c
+ * @file 01drv_lsm6ds.c
  * @author Mališa Vučinić <malisa.vucinic@inria.fr>
- * @brief This is an example on how to use the LIS2MDL driver.
+ * @brief This is an example on how to use the LSM6DS driver.
  *
- * Load this program on your board, set CALIBRATION_PROCEDURE to 1 if calibration is needed.
- * With CALIBRATION_PROCEDURE set to 0, compass heading is printed as debug output.
+ * Load this program on your board in debug mode, roll will be continuously
+ * printed.
  *
  * @copyright Inria, 2022
  *
@@ -16,12 +16,9 @@
 
 // Include BSP packages
 
-#include "lis2mdl.h"
+#include "lsm6ds.h"
 
 //=========================== defines =========================================
-
-#define CALIBRATION_PROCEDURE (0)
-#define CONST_PI              (3.14159265359f)
 
 typedef struct {
 } drv_lis2mdl_vars_t;
@@ -36,32 +33,22 @@ typedef struct {
  *  @brief The program starts executing here.
  */
 int main(void) {
-    // Init the magnetometer
-    lis2mdl_init(NULL);
-
-#if CALIBRATION_PROCEDURE
-    lis2mdl_compass_data_t offset;
-
-    lis2mdl_magnetometer_calibrate(&offset);
-    while (1) {
-        ;
-    }
-#else
-    float   heading;
     int16_t roll;
+
+    // Init the IMU
+    lsm6ds_init(NULL);
 
     while (1) {
         // processor idle until an interrupt occurs and is handled
-        if (lis2mdl_data_ready()) {
-            lis2mdl_read_heading();
-            heading = lis2mdl_last_heading() * 180 / CONST_PI;
-            printf("heading: %f\n", heading);
+        if (lsm6ds_data_ready()) {
+            lsm6ds_read_accelerometer();
+            roll = lsm6ds_last_roll();
+            printf("roll: %d\n", roll);
         }
         __WFE();
     }
     // one last instruction, doesn't do anything, it's just to have a place to put a breakpoint.
     __NOP();
-#endif
 }
 
 //=========================== functions =========================================
