+++ projects/projects-bsp-drv.emProject
@@ -586,6 +586,46 @@
     </folder>
     <configuration Name="Debug" linker_printf_fp_enabled="Float" />
   </project>
+   <project Name="01drv_lsm6ds">
+    <configuration
+      Name="Common"
+      project_dependencies="00drv_lsm6ds(drv)"
+      project_directory="01drv_lsm6ds"
+      project_type="Executable" />
+    <folder Name="Device Files">
+      <file file_name="$(DeviceHeaderFile)" />
+      <file file_name="$(DeviceCommonHeaderFile)" />
+      <file file_name="$(DeviceSystemFile)">
+        <configuration
+          Name="Common"
+          default_code_section=".init"
+          default_const_section=".init_rodata" />
+      </file>
+    </folder>
+    <folder Name="Script Files">
+      <file file_name="../../nRF/Scripts/nRF_Target.js">
+        <configuration Name="Common" file_type="Reset Script" />
+      </file>
+      <file file_name="$(DeviceLinkerScript)">
+        <configuration Name="Common" file_type="Linker Script" />
+      </file>
+      <file file_name="$(DeviceMemoryMap)">
+        <configuration Name="Common" file_type="Memory Map" />
+      </file>
+    </folder>
+    <folder Name="Source Files">
+      <configuration Name="Common" filter="c;cpp;cxx;cc;h;s;asm;inc" />
+      <file file_name="01drv_lsm6ds.c" />
+    </folder>
+    <folder Name="System Files">
+      <file file_name="$(SeggerThumbStartup)" />
+      <file file_name="$(DeviceCommonVectorsFile)" />
+      <file file_name="$(DeviceVectorsFile)">
+        <configuration Name="Common" file_type="Assembly" />
+      </file>
+    </folder>
+    <configuration Name="Debug" linker_printf_fp_enabled="Float" />
+  </project>
   <project Name="01drv_pid">
     <configuration
       Name="Common"
